requirements.txt
streamlit
geopys